add got-10k and refractor test.
