+++ test.py
@@ -3,7 +3,7 @@ import argparse
 import cv2
 import logging
 import torch
-from dataset.vot_dataset import VOTDataset
+from toolkit.datasets import get_dataset
 from utils.model_load import load_pretrain
 from models.model_builder import get_model
 from configs.config import cfg
@@ -23,27 +23,12 @@ args = parse.parse_args()
 torch.set_num_threads(1)  # use only one threads to test the real speed
 
 
-def test_tracker(data_dir, anno_file, visual=False):
-    cfg.merge_from_file(args.cfg)
-    init_log('global', logging.INFO)
-    test_dataset = VOTDataset(data_dir, anno_file)
-    # get the base model, may can be refracted.
-    if args.tracker == 'SiamRPN':
-        model_name = 'BaseSiamModel'
-    elif args.tracker == 'MetaSiamRPN':
-        model_name = 'MetaSiamModel'
-    elif args.trackers == 'GraphSiamRPN':
-        model_name = 'GraphSiamModel'
-    else:
-        raise Exception('tracker is valid')
-    base_model = get_model(model_name)
-    base_model = load_pretrain(base_model, args.snapshot).cuda().eval()
-    tracker = get_tracker(args.tracker, base_model)
+def vot_evaluate(dataset, tracker):
     tracker_name = args.tracker
     backbone_name = args.cfg.split('/')[-1].split('_')[0]
     snapshot_name = args.snapshot.split('/')[-1].split('.')[0]
     total_lost = 0
-    for v_idx, video in enumerate(test_dataset):
+    for v_idx, video in enumerate(dataset):
         if args.video != '':  # if test special video
             if video.name != args.video:
                 continue
@@ -71,7 +56,7 @@ def test_tracker(data_dir, anno_file, visual=False):
                 pred_bboxes.append(0)
 
             toc += cv2.getTickCount() - tic
-            if visual and idx > frame_count:
+            if args.vis and idx > frame_count:
                 show_double_bbox(frame, bbox, gt_bbox, idx, lost_number)
         toc /= cv2.getTickFrequency()
         result_dir = os.path.join(cfg.TRACK.RESULT_DIR, args.dataset, tracker_name, backbone_name, snapshot_name)
@@ -87,21 +72,87 @@ def test_tracker(data_dir, anno_file, visual=False):
         # log
         total_lost += lost_number
         print('[{:d}/{:d}] video: {}, time: {:.1f}s, speed: {:.1f}fps, lost_number: {:d} '.format(v_idx+1,
-                                                                                                  len(test_dataset),
+                                                                                                  len(dataset),
                                                                                                   video.name,
                                                                                                   toc, idx / toc,
                                                                                                   lost_number))
     print('total_lost: {}'.format(total_lost))
 
 
-if __name__ == '__main__':
-    if args.dataset == 'VOT2016':
-        data_dir = os.path.join(cfg.TRACK.DATA_DIR, 'VOT2016')
-        anno_file = 'VOT2016.json'
-        test_tracker(data_dir, anno_file, visual=args.vis)
-    elif args.dataset == 'VOT2018':
-        data_dir = os.path.join(cfg.TRACK.DATA_DIR, 'VOT2018')
-        anno_file = 'VOT2018.json'
-        test_tracker(data_dir, anno_file, visual=args.vis)
+def ope_evaluate(dataset, tracker):
+    tracker_name = args.tracker
+    backbone_name = args.cfg.split('/')[-1].split('_')[0]
+    snapshot_name = args.snapshot.split('/')[-1].split('.')[0]
+    for v_idx, video in enumerate(dataset):
+        if args.video != '':  # if test special video
+            if video.name != args.video:
+                continue
+        pred_bboxes = []
+        runtime = []
+        toc = 0
+        for idx, (frame, gt_bbox) in enumerate(video):
+            tic = cv2.getTickCount()
+            if idx == 0:
+                tracker.init(frame, gt_bbox)  # cx,cy,w,h
+            else:
+                track_result = tracker.track(frame)
+                bbox = track_result['bbox']  # cx,cy,w,h
+                bbox_ = [bbox[0] - bbox[2] / 2, bbox[1] - bbox[3] / 2, bbox[2], bbox[3]]  # x,y,w,h
+                gt_bbox_ = [gt_bbox[0] - gt_bbox[2] / 2, gt_bbox[1] - gt_bbox[3] / 2, gt_bbox[2], gt_bbox[3]]
+                pred_bboxes.append(bbox_)
+
+            toc += cv2.getTickCount()-tic
+            runtime.append((cv2.getTickCount()-tic)/cv2.getTickFrequency())
+            if args.vis and idx > 0:
+                show_double_bbox(frame, bbox, gt_bbox, idx, 0)
+        toc /= cv2.getTickFrequency()
+        result_dir = os.path.join(cfg.TRACK.RESULT_DIR, args.dataset, tracker_name,
+                                  backbone_name, snapshot_name, video.name)
+        if not os.path.isdir(result_dir):
+            os.makedirs(result_dir)
+        result_path = '{}/{}_001.txt'.format(result_dir, video.name)
+        runtime_path = '{}/{}_time.txt'.format(result_dir, video.name)
+        # write result
+        with open(result_path, 'w') as f:
+            for x in pred_bboxes:
+                if isinstance(x, int):
+                    f.write('{:d}\n'.format(x))
+                else:
+                    f.write(','.join(['{:.4f}'.format(i) for i in x]) + '\n')
+        # write runtime
+        with open(runtime_path, 'w') as f:
+            for time in runtime:
+                f.write('{:.6f}\n'.format(time))
+
+        # log
+        print('[{:d}/{:d}] video: {}, time: {:.1f}s, speed: {:.1f}fps'.format(v_idx + 1,
+                                                                              len(dataset),
+                                                                              video.name,
+                                                                              toc, idx / toc))
+
+
+def main():
+    cfg.merge_from_file(args.cfg)
+    init_log('global', logging.INFO)
+    # get the base model, may can be refracted.
+    if args.tracker == 'SiamRPN':
+        model_name = 'BaseSiamModel'
+    elif args.tracker == 'MetaSiamRPN':
+        model_name = 'MetaSiamModel'
+    elif args.tracker == 'GraphSiamRPN':
+        model_name = 'GraphSiamModel'
     else:
-        raise Exception('dataset invalid!')
+        raise Exception('tracker is valid')
+    base_model = get_model(model_name)
+    base_model = load_pretrain(base_model, args.snapshot).cuda().eval()
+    tracker = get_tracker(args.tracker, base_model)
+    data_dir = os.path.join(cfg.TRACK.DATA_DIR, args.dataset)
+    dataset = get_dataset(args.dataset, data_dir)
+    if args.dataset in ['VOT2016', 'VOT2018']:
+        vot_evaluate(dataset, tracker)
+    elif args.dataset == 'GOT-10k':
+        ope_evaluate(dataset, tracker)
+
+
+if __name__ == '__main__':
+    main()
